ai/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-dotenv==1.2.1
python-multipart==0.0.21
google-genai==1.59.0
pillow==12.1.0
requests==2.32.5
rembg==2.0.72
opencv-python==4.13.0.90